Quick hack to auto press enter at login
Quick hack to auto press enter upon login screen initialize. 
Allows for the first login screen to be skipped and to jump to user login screen.
For me, it allows pam_usb to complete the login authentication without a keyboard or mouse attached. (synergy then auto starts once logged in)